refactor(tests) Split test into their own files
This allows tests to run in parallel when using the key API/RemoteSigner.
