+++ integration-tests/contract-originate-token-contract.spec.ts
@@ -1,0 +1,34 @@
+import { CONFIGS } from "./config";
+import { tokenCode, tokenInit } from "./data/tokens";
+
+CONFIGS.forEach(({ lib, rpc, setup}) => {
+  const Tezos = lib;
+  describe(`Test origination of a token contract using: ${rpc}`, () => {
+
+    beforeEach(async (done) => {
+      await setup()
+      done()
+    })
+    it('originates a token contract and mints some tokens', async (done) => {
+      // TODO: Fails when using ephemeral keys
+      const op = await Tezos.contract.originate({
+        balance: "1",
+        code: tokenCode,
+        init: tokenInit(await Tezos.signer.publicKeyHash()),
+        fee: 150000,
+        storageLimit: 10000,
+        gasLimit: 400000,
+      })
+      await op.confirmation()
+      expect(op.hash).toBeDefined();
+      expect(op.includedInBlock).toBeLessThan(Number.POSITIVE_INFINITY)
+      const contract = await op.contract();
+      const opMethod = await contract.methods.mint(await Tezos.signer.publicKeyHash(), 100).send();
+
+      await opMethod.confirmation();
+      expect(op.hash).toBeDefined();
+      expect(op.includedInBlock).toBeLessThan(Number.POSITIVE_INFINITY)
+      done();
+    });
+  });
+})
