+++ integration-tests/contract-originate-vote-contract.spec.ts
@@ -1,0 +1,34 @@
+import { CONFIGS } from "./config";
+import { voteSample } from "./data/vote-contract";
+
+CONFIGS.forEach(({ lib, rpc, setup}) => {
+  const Tezos = lib;
+  describe(`Originate a voting contract using: ${rpc}`, () => {
+
+    beforeEach(async (done) => {
+      await setup()
+      done()
+    })
+  it('originates a voting contract and inits the storage', async (done) => {
+    // TODO: probably remove this as very expensive
+      const op = await Tezos.contract.originate({
+        balance: "1",
+        code: voteSample,
+        storage: {
+          mgr1: {
+            addr: await Tezos.signer.publicKeyHash(),
+            key: null,
+          },
+          mgr2: {
+            addr: await Tezos.signer.publicKeyHash(),
+            key: await Tezos.signer.publicKeyHash(),
+          },
+        }
+      })
+      await op.confirmation()
+      expect(op.hash).toBeDefined();
+      expect(op.includedInBlock).toBeLessThan(Number.POSITIVE_INFINITY)
+      done();
+    });
+  });
+})
