+++ integration-tests/contract-simple-origination.spec.ts
@@ -1,0 +1,29 @@
+import { CONFIGS } from "./config";
+
+CONFIGS.forEach(({ lib, rpc, setup}) => {
+  const Tezos = lib;
+  describe(`Originating a contract api using: ${rpc}`, () => {
+
+    beforeEach(async (done) => {
+      await setup()
+      done()
+    })
+it('Simple origination scenario', async (done) => {
+      const op = await Tezos.contract.originate({
+        balance: "1",
+        code: `parameter string;
+        storage string;
+        code {CAR;
+              PUSH string "Hello ";
+              CONCAT;
+              NIL operation; PAIR};
+        `,
+        init: `"test"`
+      })
+      await op.confirmation()
+      expect(op.hash).toBeDefined();
+      expect(op.includedInBlock).toBeLessThan(Number.POSITIVE_INFINITY)
+      done();
+    });
+  });
+})
