+++ integration-tests/contract-tx-wait-2-confirmations.spec.ts
@@ -1,0 +1,23 @@
+import { CONFIGS } from "./config";
+
+CONFIGS.forEach(({ lib, rpc, setup}) => {
+  const Tezos = lib;
+  describe(`Test tx and waiting for 2 confirmations using: ${rpc}`, () => {
+
+    beforeEach(async (done) => {
+      await setup()
+      done()
+    })
+  it('transfers 2 tez and waits for 2 confirmations', async (done) => {
+      const op = await Tezos.contract.transfer({ to: 'tz1ZfrERcALBwmAqwonRXYVQBDT9BjNjBHJu', amount: 2 })
+      await op.confirmation()
+      expect(op.includedInBlock).toBeLessThan(Number.POSITIVE_INFINITY)
+      const [first, second] = await Promise.all([op.confirmation(), op.confirmation(2)])
+      expect(second - first).toEqual(1)
+      // Retrying another time should be instant
+      const [first2, second2] = await Promise.all([op.confirmation(), op.confirmation(2)])
+      expect(second2 - first2).toEqual(1)
+      done();
+    })
+  });
+})
