+++ integration-tests/contract-unit-as-param.spec.ts
@@ -1,0 +1,26 @@
+import { CONFIGS } from "./config";
+import { depositContractCode, depositContractStorage } from "./data/deposit_contract";
+
+CONFIGS.forEach(({ lib, rpc, setup}) => {
+  const Tezos = lib;
+  describe(`Test contract with unit as params using: ${rpc}`, () => {
+
+    beforeEach(async (done) => {
+      await setup()
+      done()
+    })
+  it('originates contract and calls deposit method with unit param', async (done) => {
+      const op = await Tezos.contract.originate({
+        balance: "1",
+        code: depositContractCode,
+        init: depositContractStorage
+      })
+      const contract = await op.contract()
+
+      const operation = await contract.methods.deposit(null).send({ amount: 1, });
+      await operation.confirmation();
+      expect(operation.status).toEqual('applied')
+      done();
+    })
+  });
+})
